requirements.txt
fastapi
uvicorn
psycopg2-binary
celery
redis
requests
pytest
python-dotenv